modify: LottoNumberCollection 구성 필드 수정
